Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   scripts/variables.js

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

Date:      Tue Jan 16 19:22:43 2024 -0500

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   scripts/variables.js